+++ attachment_preview/__manifest__.py
@@ -15,9 +15,9 @@
     "assets": {
         "web._assets_primary_variables": [],
         "web.assets_backend": [
-            "attachment_preview/static/src/js/models/attachment_card/attachment_card.js",
-            "attachment_preview/static/src/js/attachmentPreviewWidget.js",
-            "attachment_preview/static/src/js/components/chatter/chatter.js",
+            "attachment_preview/static/src/js/models/attachment_card/attachment_card.esm.js",
+            "attachment_preview/static/src/js/attachmentPreviewWidget.esm.js",
+            "attachment_preview/static/src/js/components/chatter/chatter.esm.js",
             "attachment_preview/static/src/scss/attachment_preview.scss",
         ],
         "web.assets_frontend": [],
