+++ attachment_preview/static/src/js/attachmentPreviewWidget.esm.js
@@ -1,6 +1,5 @@
 /** @odoo-module */
 import Widget from "web.Widget";
-import {_t} from "web.core";
 
 var active_attachment_index = 0;
 var is_first_click = true;
@@ -17,7 +16,7 @@ var AttachmentPreviewWidget = Widget.extend({
     },
 
     start: function () {
-        // first_click = true;
+        // First_click = true;
         var res = this._super.apply(this, arguments);
         this.$overlay = $(".attachment_preview_overlay");
         this.$iframe = $(".attachment_preview_iframe");
@@ -102,7 +101,9 @@ var AttachmentPreviewWidget = Widget.extend({
 
         if (is_first_click) {
             for (let i = 0; i < this.attachments.length; i++) {
-                if (parseInt(this.attachments[i].id) === this.active_attachment_id) {
+                if (
+                    parseInt(this.attachments[i].id, 10) === this.active_attachment_id
+                ) {
                     active_attachment_index = i;
                     is_first_click = false;
                 }
